refactor: Observer design pattern was implemented
The Transaction class now implements a Subject interface while the
Reminder class acts as an Observer in the Observer design pattern. Data
access object classes now handle database access and data manipulation
that is used by the main classes.

In addition, the "Check Reminder" button needs to be re-implemented, as
this will function to display employee reminders in this pattern.